fix(st): 收口 L0 推送后的 clang-tidy error 级三处
目的：L0 推送后 Linux CI 的 E2 clang-tidy 门报三处 bugprone error 级违例（misplaced-widening-cast、branch-clone ×2），按门禁零违例要求修复；readability 级警告按既有口径保持警告不阻断。

修改内容：
- lexer.h：十六进制位值计算改为无窄化的 u64 运算（语义等价）
- pins.h：TON/TOF/TP 共享引脚表的三个相同分支合并 fallthrough
- sema.h：valid_arith 中 multiply/divide 相同分支合并

影响范围：纯词法级重写，字节码输出逐位不变（确定性锚点哈希 8149969011563308332 复验一致），本地 ST 测试全绿。跨平台确定性已由本次推送的 gcc 作业证实：Linux 53/53 通过含锚点哈希门，ARM64 交叉编译通过。

AI 生成声明：本提交由 Claude (AI) 在 CI 门禁失败日志下生成。

+++ core/st/lexer.h
@@ -366,7 +366,8 @@ private:
                 if(lower < 'a' || lower > 'f') {
                     break;
                 }
-                digit = static_cast<std::uint64_t>(lower - 'a' + 10);
+                digit = static_cast<std::uint64_t>(lower) -
+                        static_cast<std::uint64_t>('a') + 10;
             } else {
                 break;
             }
